Add comment about current directory in path
I have known about the security risks with having current directory in
my path, but comfortable with the risk/benefit trade off on my single
user machines.  However, now that this is part of a public repository, I
have added a comment about the security risk and commented out the line
by default.